Here is the real elimination of the segmentation-fault on shutdown after restoring modified internal functions.
This problem was fully described here http://pecl.php.net/bugs/bug.php?id=11632
but I see this bug occurs only after the second request to the test-page through fascgi (on the same worker process) or results in crash in cli mode for the each test execution.

The solution is to remove all user-defined functions out of the executor's global function table in a request shutdown function.